Make reclineview work with GET requests again
The city of Zurich portal doesn't allow any other requests than GET and
therefore recline view needs to be updated to get it's data in a
different way. The datastore_search API works with GET and POST so there
is no problem in changing the method.

With GET there is no need to stringify the data JSON.